+++ src/main/java/com/github/davidmoten/rtree/geometry/ListPair.java
@@ -14,8 +14,8 @@ public final class ListPair<T extends HasGeometry> {
     private final Group<T> group2;
     // these non-final variable mean that this class is not thread-safe
     // because access to them is not synchronized
-    private float areaSum = -1;
-    private final float marginSum;
+    private double areaSum = -1;
+    private final double marginSum;
 
     public ListPair(List<T> list1, List<T> list2) {
         this.group1 = new Group<T>(list1);
@@ -31,13 +31,13 @@ public final class ListPair<T extends HasGeometry> {
         return group2;
     }
 
-    public float areaSum() {
+    public double areaSum() {
         if (areaSum == -1)
             areaSum = group1.geometry().mbr().area() + group2.geometry().mbr().area();
         return areaSum;
     }
 
-    public float marginSum() {
+    public double marginSum() {
         return marginSum;
     }
